requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
cachelib==0.9.0
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
Flask-Session==0.4.0
frozenlist==1.3.3
idna==3.4
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.3
pyee==9.0.4
python-dotenv==0.21.0
slackclient==2.9.4
slackeventsapi==3.0.1
typing_extensions==4.4.0
Werkzeug==2.2.2
yarl==1.8.2
zipp==3.11.0
